- remove unnecessary files - rename VO files - create a new package for kakaopay service
