+++ src/main/java/withus/ex/vo/AmountVO.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @Setter
 @Getter
 @ToString
-public class Amount {
+public class AmountVO {
 	
     private int total; // 총 결제 금액
     private int tax_free; // 비과세 금액
