+++ src/main/java/withus/ex/vo/CanceledAmountVO.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @Setter
 @Getter
 @ToString
-public class CanceledAmount {
+public class CanceledAmountVO {
 	
     private int total; // 취소된 전체 누적 금액
     private int tax_free; // 취소된 비과세 누적 금액
