+++ src/main/java/withus/ex/vo/OrderVO.java
@@ -1,7 +1,5 @@
 package withus.ex.vo;
 
-import java.util.List;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
