--- src/main/resources/mappers/ProductMapper.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-
-<mapper namespace="withus.ex.mapper.ProductMapper">
-
-	<select id="selectList" resultMap="productListMap">
-	<![CDATA[
-		select * from withus_product
-	]]>
-	</select>
-<!-- 		<![CDATA[
-			select * from withus_product p, withus_product_img pi
-			where p.product_number = pi.product_number and withus_product = #{productNumber}
-		]]> -->
-	<!-- 	<![CDATA[
-			select * from withus_product_img
-		]]> -->
-		
-	<select id="selectImgList" resultMap="productImgMap">
-	 	select * from withus_product_img
-	</select>
-	<select id="selectDetail" resultType="withus.ex.vo.ProductVO">
-	<![CDATA[
-			select p.product_number,
-			 pi.product_img_number, 
-			 p.product_name, 
-			 p.main_img, 
-			 p.price, 
-			 p.order_quantity, 
-			 p.explanation, 
-			 pi.image_type, 
-			 pi.file_name 
-			 from withus_product p, withus_product_img pi 
-			 where p.product_number = pi.product_number and p.product_number = #{product_number}
-		]]>
-	</select>
-
-   <!-- DB테이블 하고 VO객체를 매핑, 조인이니까 resultMap사용함 -->
-   <resultMap id="productListMap" type="withus.ex.vo.ProductVO">
-      <result property="product_number" column="product_number" /> <!-- userVO의 속성인 getUsername()하고 DB테이블의 컬럼인 username이 매핑됨 -->
-      <result property="product_img_number" column="product_img_number" />
-      <result property="product_name" column="product_name" />
-      <result property="main_img" column="product_img" />
-      <result property="price" column="price" />
-      <result property="order_quantity" column="order_quantity" />
-      <result property="explanation" column="explanation" />
-	<collection property="imgList" resultMap="productImgMap" />
-   
-   </resultMap>   
-    
-      <!-- DB테이블 하고 VO객체를 매핑, 조인이니까 resultMap사용함 -->
-   <resultMap id="productImgMap" type="withus.ex.vo.ProductImgVO">
-      <result property="product_img_number" column="product_img_number" /> <!-- userVO의 속성인 getUsername()하고 DB테이블의 컬럼인 username이 매핑됨 -->
-      <result property="product_number" column="product_number" />
-      <result property="image_type" column="image_type" />
-      <result property="file_name" column="file_name" />
-
-  
-   </resultMap>
-<!--    
-   <select id="getUser" resultMap="userMap"> 부모매퍼인터페이스를 구현한거임
-      select * from users , authorities
-      where users.username = authorities.username and users.username = #{username}
-      
-   </select>
-
-<resultMap id="userMap" type="withus.ex.vo.productListVO">
-      <result property="userNumber" column="USER_NUMBER" /> userVO의 속성인 getUsername()하고 DB테이블의 컬럼인 username이 매핑됨
-      <result property="product_img_number" column="product_img_number" />
-      <collection property="authList" resultMap="authMap"></collection> 
-   </resultMap>
- -->
-
-
-</mapper>
